backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
feedparser
requests
psycopg2-binary
python-dateutil
tenacity
pytest
pytest-asyncio
factory-boy
email-validator